Add files via upload
I've made adjustments so that the project fully builds and renders the windows, currently working on testimonials. There was an issue with the lack of implementation files for the ui's for testimonials, about, and reports. I'm currently making adjustments to add them in right now, and linking the menus to open the widgets and doing some backend stuff for functionality. There is also a folder inside the project called "TextFiles" for where the usernames and passwords, as well as testimonials will be stored. There should not be any issues with saving testimonies but LET ME KNOW IF THERE IS (It'll launch a dialog that says failure to open file). I'll most likely upload an updated version sometime tonight. Remember to delete the pro.user file, I forgot to do it on this upload.